requirements.txt
Flask==2.2.5
pymysql
boto3
python-dotenv